+++ src/index.js
@@ -1,17 +1,35 @@
 import GithubFeed from './components/GithubFeed.vue'
-import GithubEvent from './components/GithubEvent.vue'
 import PushEvent from './components/events/PushEvent.vue'
+import PullRequestEvent from './components/events/PullRequestEvent.vue'
+import CreateEvent from './components/events/CreateEvent.vue'
+import DeleteEvent from './components/events/DeleteEvent.vue'
+import WatchEvent from './components/events/WatchEvent.vue'
+import IssuesEvent from './components/events/IssuesEvent.vue'
+import IssueCommentEvent from './components/events/IssueCommentEvent.vue'
+import ForkEvent from './components/events/ForkEvent.vue'
 
 export function install (Vue) {
   Vue.component('github-feed', GithubFeed)
-  Vue.component('github-event', GithubEvent)
   Vue.component('push-event', PushEvent)
+  Vue.component('pull-request-event', PullRequestEvent)
+  Vue.component('create-event', CreateEvent)
+  Vue.component('watch-event', WatchEvent)
+  Vue.component('delete-event', DeleteEvent)
+  Vue.component('issues-event', IssuesEvent)
+  Vue.component('issue-comment-event', IssueCommentEvent)
+  Vue.component('fork-event', ForkEvent)
 }
 
 export {
   GithubFeed,
-  GithubEvent,
   PushEvent,
+  PullRequestEvent,
+  CreateEvent,
+  WatchEvent,
+  DeleteEvent,
+  IssuesEvent,
+  IssueCommentEvent,
+  ForkEvent,
 }
 
 const plugin = {